Respect last window state
This updates the main window's initialisation code to remember the window state from the previous time the app was used, using the `electron-window-state` helper.

Rather than maximising every time the application starts, it will instead start at 1000 by 800, and subsequently open in whichever size the user last used it in.

This fixes #47.